feat: serviço e controller do Associado , save , update , findById, delete

+++ src/main/java/com/br/VotaCoop/domain/repository/AssociadoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssociadoRepository extends JpaRepository<Associado, Integer> {
+public interface AssociadoRepository extends JpaRepository<Associado, Long> {
 }
