+++ src/main/java/com/br/VotaCoop/domain/repository/PautaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PautaRepository extends JpaRepository<Pauta, Integer> {
+public interface PautaRepository extends JpaRepository<Pauta, Long> {
 }
