+++ src/main/java/com/br/VotaCoop/domain/repository/VotoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VotoRepository extends JpaRepository<Voto, Integer> {
+public interface VotoRepository extends JpaRepository<Voto, Long> {
 }
